libdrgn: index attributes related to DWARF blocks
In preparation to handle function parameters and stack variables, we'll
need to be able to discover the location information associated with them.

This commit adds indexing for the following DWARF attributes:
- DW_AT_low_pc
- DW_AT_high_pc
- DW_AT_abstract_origin

Since the ref_addr form of DW_AT_abstract_origin uses the compilation
unit offset size, we'll track that too.