libdrgn: extend drgn_object to describe function parameters and stack vars
Now that we can locate the debuginfo for any given function, we can
describe the type and location for the function parameters and stack
variables for a given pc.

This works by enumerating the DW_TAG_formal_parameter tags for the
function and DW_TAG_variables for every scope that contains the pc (but
within the function).  We assemble a frame symbol that contains the name
and the Dwarf_Die that describes it.  When the object is created, it's
marked as needing evaluation as a stack object.  When that evaluation
happens, we look up the debuginfo and assemble the address (or value)
for the object, and clone the initial object to populate it with the
real values.

Since the debuginfo and core dumps are imperfect, we introduce several
new error types:
- DRGN_ERROR_VAR_OPTIMIZED_OUT -- No debuginfo found for the object at all
- DRGN_ERROR_VAR_LOCATION_UNAVAILABLE -- Incomplete debuginfo found, which
  doesn't cover the current pc
- DRGN_ERROR_VAR_VALUE_UNAVAILABLE -- The debuginfo is complete but the
  execution state can't be located

Each can be raised when the location is evaluated, so that the object
can exist with the type, size, etc regardless of whether the debuginfo
is complete.

The python code can lookup the parameters and variables by name or index,
and the stack frame object can be subscripted as a dictionary.  Shadowing
of variable and parameter names are handled by evaluating the scopes from
the inside out, and keeping that ordering of evaluation in the vector
of symbols.  A search of the vector will match the variable in the
innermost scope first.  Likewise, a variable that shadows a parameter
will take precedence.  It is possible to enumerate all the variables so that
even shadowed variables can be located, just not directly by name.